refactor to courses
