+++ app/src/main/AndroidManifest.xml
@@ -59,7 +59,7 @@
             android:name=".ui.userprofile.BecomeMemberActivity"
             android:theme="@style/AppTheme" />
         <activity
-            android:name=".ui.course.CourseProgressActivity"
+            android:name=".ui.courses.CourseProgressActivity"
             android:theme="@style/AppTheme" />
         <activity
             android:name=".ui.news.NewsDetailActivity"
