+++ app/src/main/java/org/ole/planet/myplanet/ui/courses/AdapterCourses.kt
@@ -1,9 +1,8 @@
-package org.ole.planet.myplanet.ui.course
+package org.ole.planet.myplanet.ui.courses
 
 import android.content.Context
 import android.os.Build
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.MotionEvent
 import android.view.View
