+++ app/src/main/java/org/ole/planet/myplanet/ui/courses/AdapterMyProgress.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.course
+package org.ole.planet.myplanet.ui.courses
 
 import android.content.Context
 import android.content.Intent
