+++ app/src/main/java/org/ole/planet/myplanet/ui/courses/CourseProgressActivity.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.course
+package org.ole.planet.myplanet.ui.courses
 
 import android.os.Bundle
 import androidx.recyclerview.widget.GridLayoutManager
