+++ app/src/main/java/org/ole/planet/myplanet/ui/courses/MyProgressFragment.kt
@@ -1,4 +1,4 @@
-package org.ole.planet.myplanet.ui.course
+package org.ole.planet.myplanet.ui.courses
 
 import android.os.Bundle
 import android.view.LayoutInflater
