+++ app/src/main/java/org/ole/planet/myplanet/ui/dashboard/DashboardActivity.kt
@@ -42,7 +42,7 @@ import org.ole.planet.myplanet.service.UserProfileDbHandler
 import org.ole.planet.myplanet.ui.SettingActivity
 import org.ole.planet.myplanet.ui.chat.ChatHistoryListFragment
 import org.ole.planet.myplanet.ui.community.CommunityTabFragment
-import org.ole.planet.myplanet.ui.course.CourseFragment
+import org.ole.planet.myplanet.ui.courses.CoursesFragment
 import org.ole.planet.myplanet.ui.feedback.FeedbackListFragment
 import org.ole.planet.myplanet.ui.resources.ResourceDetailFragment
 import org.ole.planet.myplanet.ui.resources.ResourcesFragment
@@ -284,11 +284,11 @@ class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, B
             R.string.menu_library -> openCallFragment(ResourcesFragment())
             R.string.menu_meetups -> {}
             R.string.menu_surveys -> openCallFragment(SurveyFragment())
-            R.string.menu_courses -> openCallFragment(CourseFragment())
+            R.string.menu_courses -> openCallFragment(CoursesFragment())
             R.string.menu_community -> openCallFragment(CommunityTabFragment())
             R.string.txt_myLibrary -> openMyFragment(ResourcesFragment())
             R.string.team -> openMyFragment(TeamFragment())
-            R.string.txt_myCourses -> openMyFragment(CourseFragment())
+            R.string.txt_myCourses -> openMyFragment(CoursesFragment())
             R.string.enterprises -> openEnterpriseFragment()
             R.string.menu_logout -> logout()
             else -> openCallFragment(BellDashboardFragment())
@@ -400,10 +400,10 @@ class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, B
                 openCallFragment(ResourcesFragment())
             }
             R.id.menu_courses -> {
-                openCallFragment(CourseFragment())
+                openCallFragment(CoursesFragment())
             }
             R.id.menu_mycourses -> {
-                openMyFragment(CourseFragment())
+                openMyFragment(CoursesFragment())
             }
             R.id.menu_mylibrary -> {
                 openMyFragment(ResourcesFragment())
