+++ app/src/main/java/org/ole/planet/myplanet/ui/dashboard/DashboardActivity.kt.lite
@@ -31,6 +31,7 @@ import com.mikepenz.materialdrawer.model.PrimaryDrawerItem
 import com.mikepenz.materialdrawer.model.interfaces.IDrawerItem
 import com.mikepenz.materialdrawer.model.interfaces.Nameable
 import org.ole.planet.myplanet.R
+//import org.ole.planet.myplanet.base.BaseContainerFragment
 import org.ole.planet.myplanet.callback.OnHomeItemClickListener
 import org.ole.planet.myplanet.databinding.ActivityDashboardBinding
 import org.ole.planet.myplanet.databinding.CustomTabBinding
@@ -41,7 +42,7 @@ import org.ole.planet.myplanet.service.UserProfileDbHandler
 import org.ole.planet.myplanet.ui.SettingActivity
 import org.ole.planet.myplanet.ui.chat.ChatHistoryListFragment
 import org.ole.planet.myplanet.ui.community.CommunityTabFragment
-import org.ole.planet.myplanet.ui.course.CourseFragment
+import org.ole.planet.myplanet.ui.courses.CoursesFragment
 import org.ole.planet.myplanet.ui.feedback.FeedbackListFragment
 import org.ole.planet.myplanet.ui.resources.ResourceDetailFragment
 import org.ole.planet.myplanet.ui.resources.ResourcesFragment
@@ -56,7 +57,6 @@ import org.ole.planet.myplanet.utilities.Constants.showBetaFeature
 import org.ole.planet.myplanet.utilities.KeyboardUtils.setupUI
 import org.ole.planet.myplanet.utilities.LocaleHelper
 import org.ole.planet.myplanet.utilities.Utilities
-//import org.ole.planet.myplanet.base.BaseContainerFragment
 
 class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, BottomNavigationView.OnNavigationItemSelectedListener {
     private lateinit var activityDashboardBinding: ActivityDashboardBinding
@@ -284,11 +284,11 @@ class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, B
             R.string.menu_library -> openCallFragment(ResourcesFragment())
             R.string.menu_meetups -> {}
             R.string.menu_surveys -> openCallFragment(SurveyFragment())
-            R.string.menu_courses -> openCallFragment(CourseFragment())
+            R.string.menu_courses -> openCallFragment(CoursesFragment())
             R.string.menu_community -> openCallFragment(CommunityTabFragment())
             R.string.txt_myLibrary -> openMyFragment(ResourcesFragment())
             R.string.team -> openMyFragment(TeamFragment())
-            R.string.txt_myCourses -> openMyFragment(CourseFragment())
+            R.string.txt_myCourses -> openMyFragment(CoursesFragment())
             R.string.enterprises -> openEnterpriseFragment()
             R.string.menu_logout -> logout()
             else -> openCallFragment(BellDashboardFragment())
@@ -400,10 +400,10 @@ class DashboardActivity : DashboardElementActivity(), OnHomeItemClickListener, B
                 openCallFragment(ResourcesFragment())
             }
             R.id.menu_courses -> {
-                openCallFragment(CourseFragment())
+                openCallFragment(CoursesFragment())
             }
             R.id.menu_mycourses -> {
-                openMyFragment(CourseFragment())
+                openMyFragment(CoursesFragment())
             }
             R.id.menu_mylibrary -> {
                 openMyFragment(ResourcesFragment())
