+++ app/src/main/res/layout/activity_course_progress.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.course.CourseProgressActivity">
+    tools:context=".ui.courses.CourseProgressActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
