+++ app/src/main/res/layout/fragment_course_detail.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.course.CourseDetailFragment">
+    tools:context=".ui.courses.CourseDetailFragment">
 
     <androidx.core.widget.NestedScrollView
         android:layout_width="match_parent"
