+++ app/src/main/res/layout/fragment_course_step.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.course.CourseStepFragment">
+    tools:context=".ui.courses.CourseStepFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
