+++ app/src/main/res/layout/fragment_my_progress.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.course.MyProgressFragment">
+    tools:context=".ui.courses.MyProgressFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_myprogress"
